♻️ simplify edit user to restrict it to editing roles only
